Move GADT constraint code to a separate file
Also rename the classes to better reflect their role,
and document and reorder definitions to make more sense.
